+++ src/ServiceManager.php
@@ -141,6 +141,24 @@ class ServiceManager implements ServiceLocatorInterface
     }
 
     /**
+     * Implemented for backwards compatibility with previous plugin managers only.
+     *
+     * Returns the creation context.
+     *
+     * @deprecated since 3.0.0. Factories using 3.0 should use the container
+     *     instance passed to the factory instead.
+     * @return ContainerInterface
+     */
+    public function getServiceLocator()
+    {
+        trigger_error(sprintf(
+            'Usage of %s is deprecated since v3.0.0; please use the container passed to the factory instead',
+            __METHOD__
+        ), E_USER_DEPRECATED);
+        return $this->creationContext;
+    }
+
+    /**
      * {@inheritDoc}
      */
     public function get($name)
@@ -210,7 +228,7 @@ class ServiceManager implements ServiceLocatorInterface
 
         // Check abstract factories
         foreach ($this->abstractFactories as $abstractFactory) {
-            if ($abstractFactory->canCreateServiceWithName($this, $name)) {
+            if ($abstractFactory->canCreate($this, $name)) {
                 return true;
             }
         }
@@ -590,7 +608,7 @@ class ServiceManager implements ServiceLocatorInterface
 
         // Check abstract factories
         foreach ($this->abstractFactories as $abstractFactory) {
-            if ($abstractFactory->canCreateServiceWithName($this, $name)) {
+            if ($abstractFactory->canCreate($this, $name)) {
                 return $abstractFactory;
             }
         }