+++ test/TestAsset/V2ValidationPluginManager.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * Zend Framework (http://framework.zend.com/)
+ *
+ * @link      http://github.com/zendframework/zf2 for the canonical source repository
+ * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   http://framework.zend.com/license/new-bsd New BSD License
+ */
+
+namespace ZendTest\ServiceManager\TestAsset;
+
+use RuntimeException;
+use Zend\ServiceManager\AbstractPluginManager;
+
+class V2ValidationPluginManager extends AbstractPluginManager
+{
+    public $assertion;
+
+    public function validatePlugin($plugin)
+    {
+        if (! is_callable($this->assertion)) {
+            throw new RuntimeException(sprintf(
+                '%s requires a callable $assertion property; not currently set',
+                __CLASS__
+            ));
+        }
+
+        call_user_func($this->assertion, $plugin);
+    }
+}